pyproject.toml
[tool.poetry]
name = "sops-kdf"
version = "0.1.0"
description = "Deterministic SOPS keys"
authors = ["You"]
packages = [{include = "sops_kdf"}]

[tool.poetry.dependencies]
python = "^3.10"
bech32 = "^1.2.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
sops-kdf-hook = "sops_kdf.main:run"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
